add data point logic
